refactor: isolate Business Profile transport
